Clean up remainders in discriminant creation
The code had used ceiling division to calculate remainders, and changed
the sign of the numerator where required.  This was confusing and made
the code bug-prone.  Instead, use floor division where it is needed and
ceiling division where it is needed.  This makes the code clearer and
easier to read.